refType: test fixes for early node versions.
Drop node 0.8 from the tests. With --use_strict activated, node 0.8
fails to run mocha tests with the following message: "SyntaxError: In
strict mode code, functions can only be declared at top level or
immediately within another function." This happens even with an empty
dir and a stub test file.

In node 0.10, using getOwnPropertyDescriptor on the name property of
the new object returned by refType gives an undefined (i.e. no
property defined for it). Also, on Windows with node 0.10, the name
property doesn't have the "*" appended to the name either.

In node 0.12, we also skip testing the prior state of the read
property since it defaults to true in that version.